Retry GitHub download failures
